Move WidgetState to widget package
