+++ android/app/src/main/kotlin/net/mullvad/mullvadvpn/applist/ListItemData.kt
@@ -3,7 +3,7 @@ package net.mullvad.mullvadvpn.applist
 import androidx.annotation.DrawableRes
 import androidx.annotation.IntDef
 import androidx.annotation.StringRes
-import net.mullvad.mullvadvpn.model.WidgetState
+import net.mullvad.mullvadvpn.ui.widget.WidgetState
 
 data class ListItemData
 private constructor(
