+++ android/app/src/main/kotlin/net/mullvad/mullvadvpn/ui/widget/WidgetState.kt
@@ -1,4 +1,4 @@
-package net.mullvad.mullvadvpn.model
+package net.mullvad.mullvadvpn.ui.widget
 
 import androidx.annotation.DrawableRes
